+++ client/src/components/diagram-editor/components/DiagramControls.tsx
